requirements.txt
pycollada
open3d
numpy
scipy